+++ app/policies/group_policy.rb
@@ -48,11 +48,9 @@ class GroupPolicy < ApplicationPolicy
     update?
   end
 
-  # true if the current user is the owner of the group
-  # true if the current user is a active member of the group
+  # true if the current user is an active member of the group
   def add_slot?
     return false unless current_user?
-    return true if group.owner == current_user
     return true if current_user.active_member? group
     false
   end
@@ -102,5 +100,12 @@ class GroupPolicy < ApplicationPolicy
 
   def member_settings?
     show?
+  end
+
+  # true if the group is public and belongs to
+  # the crawler source making the request
+  def global_group?
+    return true if group.public? && group.owner == current_user
+    false
   end
 end