+++ spec/routing/v1/groups_routing_spec.rb
@@ -77,6 +77,11 @@ RSpec.describe V1::GroupsController, type: :routing do
                      format: :json)
     end
 
+    it "routes to #global_group" do
+      expect(post: "/v1/groups/global_group")
+        .to route_to("v1/groups#global_group", format: :json)
+    end
+
     it "does't route to #new" do
       expect(get: "/v1/groups/new")
         .not_to route_to("v1/groups#new", format: :json)